Audit cleanup (#109)
* split srl audit into multiple files

* edit imports

* edit wording

* show url in srl audit run

* assert url in audit run output

* move format_problem to utils

* remove dest

* Enhance doc comment and rename function
