+++ srl/commands/list_.py
@@ -1,7 +1,11 @@
 from rich.console import Console
 from rich.panel import Panel
-from srl.utils import today
-from srl.commands.audit import get_current_audit, random_audit, get_last_audit_date
+from srl.utils import today, format_problem
+from srl.commands.audit.utils import (
+    get_current_audit,
+    random_audit,
+    get_last_audit_date,
+)
 from datetime import datetime, timedelta
 import random
 from srl.storage import (
@@ -34,7 +38,7 @@ def add_subparser(subparsers):
 
 def handle(args, console: Console):
     if should_audit() and not get_current_audit():
-        problem, problem_url = random_audit(with_url=True)
+        problem, problem_url = random_audit()
         if problem:
             console.print("[bold red]You have been randomly audited![/bold red]")
             display = format_problem(problem, problem_url)
@@ -85,14 +89,6 @@ def should_audit():
     except (ValueError, TypeError):
         probability = 0.1
     return random.random() < probability
-
-
-def format_problem(problem: str, problem_url: str | None):
-    """Returns "problem (url)" if url is present, otherwise "problem" """
-    if problem_url:
-        return f"{problem} ([blue]{problem_url}[/blue])"
-
-    return problem
 
 
 def get_due_problems(limit: int | None = None) -> list[tuple[str, str]]:
